@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,14 +39,14 @@ dependencies:
   cupertino_icons: ^1.0.2
   get_it: ^7.2.0
   provider: ^6.0.3
-  connectivity: ^3.0.3
+  connectivity_plus: ^4.0.0
   dio: ^4.0.0-prev2
   http: ^0.13.1
   freezed: ^2.0.3+1
   shared_preferences: ^2.0.5
   flutter_svg: ^0.22.0
   dropdown_button2: ^2.0.0
-  carousel_slider: ^4.0.0
+  carousel_slider: ^5.0.0
   image_network:
 
 dev_dependencies:
@@ -72,8 +72,8 @@ flutter_icons:
   web:
     generate: true
     image_path: "assets/images/logo.png"
-    background_color: "#FFFFFF"
-    theme_color: "#FFFFFF"
+    background_color: "#000000"
+    theme_color: "#000000"
   windows:
     generate: true
     image_path: "assets/images/logo.png"
