@@ -1,8 +1,8 @@
 class OTPLoginModel {
-  String email;
-  String phone;
-  String code;
-  String deviceToken;
+  String? email;
+  String? phone;
+  String? code;
+  String? deviceToken;
   OTPLoginModel({this.email, this.phone , this.code , this.deviceToken});
 
   OTPLoginModel.fromJson(Map<String, dynamic> json) {
@@ -13,11 +13,11 @@ class OTPLoginModel {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-  email==null?print(""):  data['email'] = this.email;
-    phone==null?print(""): data['phone'] = this.phone;
-    data['code'] = this.code;
-    data['deviceToken'] = this.deviceToken;
+    final Map<String, dynamic> data = <String, dynamic>{};
+  email==null?print(""):  data['email'] = email;
+    phone==null?print(""): data['phone'] = phone;
+    data['code'] = code;
+    data['deviceToken'] = deviceToken;
     return data;
   }
 }
