@@ -1,12 +1,12 @@
 import 'package:dio/dio.dart';
 
 class ApiResponse {
-  final Response response;
+  final Response? response;
   final dynamic error;
 
   ApiResponse(this.response, this.error);
 
-  ApiResponse.withError(dynamic errorValue)
+  ApiResponse.withError(dynamic errorValue )
       : response = null,
         error = errorValue;
 
