@@ -1,6 +1,6 @@
 class CategoriesModel {
-  int id;
-  String name;
+  int? id;
+  String? name;
 
   CategoriesModel({
     this.id,
@@ -13,9 +13,9 @@ class CategoriesModel {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
-    data['id'] = this.id;
-    data['name'] = this.name;
+    final Map<String, dynamic> data = <String, dynamic>{};
+    data['id'] = id;
+    data['name'] = name;
 
     return data;
   }
