@@ -1,4 +1,3 @@
-import 'package:dio/dio.dart';
 
 import '../../utill/app_constants.dart';
 import '../datasource/remote/dio/dio_client.dart';
@@ -6,21 +5,21 @@ import '../datasource/remote/exception/api_error_handler.dart';
 import '../model/response/base/api_response.dart';
 
 class CategoriesRepo {
-  final DioClient dioClient;
+  final DioClient? dioClient;
 
   CategoriesRepo({this.dioClient});
   Future<ApiResponse> getCategoriesList() async {
     try {
-      final response = await dioClient.get(AppConstants.CategoriesListURL);
+      final response = await dioClient!.get(AppConstants.CategoriesListURL);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
     }
   }
 
-  Future<ApiResponse> getSubCategoriesListById({String id}) async {
+  Future<ApiResponse> getSubCategoriesListById({String? id}) async {
     try {
-      final response = await dioClient.get(AppConstants.SubCategoriesListURL+id);
+      final response = await dioClient!.get(AppConstants.SubCategoriesListURL+id!);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
