@@ -6,14 +6,14 @@ import '../datasource/remote/exception/api_error_handler.dart';
 import '../model/response/base/api_response.dart';
 
 class ConstantsRepo {
-  final DioClient dioClient;
+  final DioClient? dioClient;
 
   ConstantsRepo({this.dioClient});
 
 
   Future<ApiResponse> getTermPolicy() async {
     try {
-      final response = await dioClient.get(AppConstants.TermPolicyURL);
+      final response = await dioClient!.get(AppConstants.TermPolicyURL);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
@@ -21,7 +21,7 @@ class ConstantsRepo {
   }
   Future<ApiResponse> getAboutUsModel() async {
     try {
-      final response = await dioClient.get(AppConstants.AboutUsURL);
+      final response = await dioClient!.get(AppConstants.AboutUsURL);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
@@ -30,7 +30,7 @@ class ConstantsRepo {
 
   Future<ApiResponse> getWorkNatureList() async {
     try {
-      final response = await dioClient.get(AppConstants.WorkNatureListURL);
+      final response = await dioClient!.get(AppConstants.WorkNatureListURL);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
@@ -38,9 +38,10 @@ class ConstantsRepo {
   }
 
 
-  Future<ApiResponse> technicalSupport({String fullName , String email , String phoneNumber,String message,} ) async {
+  Future<ApiResponse> technicalSupport({String? fullName ,
+    String? email , String? phoneNumber,String? message,} ) async {
     try {
-      Response response = await dioClient.post(
+      Response response = await dioClient!.post(
         AppConstants.TechnicalSupportURL,
         data:'{"fullName": "$fullName","email": "$email","phoneNumber": "$phoneNumber","message": "$message"}'
         ,   options: Options(headers: {'accept': '*/*',
