@@ -1,4 +1,3 @@
-import 'package:dio/dio.dart';
 
 import '../../utill/app_constants.dart';
 import '../datasource/remote/dio/dio_client.dart';
@@ -6,12 +5,12 @@ import '../datasource/remote/exception/api_error_handler.dart';
 import '../model/response/base/api_response.dart';
 
 class ProductRepo {
-  final DioClient dioClient;
+  final DioClient? dioClient;
 
   ProductRepo({this.dioClient});
   Future<ApiResponse> getMostPopularProductList() async {
     try {
-      final response = await dioClient.get(AppConstants.GetMostPopularProductURL);
+      final response = await dioClient!.get(AppConstants.GetMostPopularProductURL);
       return ApiResponse.withSuccess(response);
     } catch (e) {
       return ApiResponse.withError(ApiErrorHandler.getMessage(e));
