@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:connectivity/connectivity.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 
 
@@ -9,22 +9,22 @@ class NetworkInfo {
   NetworkInfo(this.connectivity);
 
    Future<bool> get isConnected async {
-    ConnectivityResult _result = await connectivity.checkConnectivity();
-    return _result != ConnectivityResult.none;
+    ConnectivityResult result = await connectivity.checkConnectivity();
+    return result != ConnectivityResult.none;
   }
 
   static void checkConnectivity(BuildContext context) {
-    bool _firstTime = true;
+    bool? firstTime = true;
     Connectivity().onConnectivityChanged.listen((ConnectivityResult result) async {
-      if(!_firstTime) {
+      if(!firstTime!) {
         //bool isNotConnected = result != ConnectivityResult.wifi && result != ConnectivityResult.mobile;
-        bool isNotConnected;
+        bool? isNotConnected;
         if(result == ConnectivityResult.none) {
           isNotConnected = true;
         }else {
-          isNotConnected = !await _updateConnectivityStatus();
+          isNotConnected = await _updateConnectivityStatus();
         }
-        isNotConnected ? SizedBox() : ScaffoldMessenger.of(context).hideCurrentSnackBar();
+        isNotConnected! ? const SizedBox() : ScaffoldMessenger.of(context).hideCurrentSnackBar();
         ScaffoldMessenger.of(context).showSnackBar(SnackBar(
           backgroundColor: isNotConnected ? Colors.red : Colors.green,
           duration: Duration(seconds: isNotConnected ? 6000 : 3),
@@ -34,20 +34,20 @@ class NetworkInfo {
           ),
         ));
       }
-      _firstTime = false;
+      firstTime = false;
     });
   }
 
-  static Future<bool> _updateConnectivityStatus() async {
-     bool _isConnected;
+  static Future<bool?> _updateConnectivityStatus() async {
+     bool? isConnected;
      try {
-       final List<InternetAddress> _result = await InternetAddress.lookup('google.com');
-       if(_result.isNotEmpty && _result[0].rawAddress.isNotEmpty) {
-         _isConnected = true;
+       final List<InternetAddress> result = await InternetAddress.lookup('google.com');
+       if(result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
+         isConnected = true;
        }
      }catch(e) {
-       _isConnected = false;
+       isConnected = false;
      }
-     return _isConnected;
+     return isConnected;
   }
 }
