@@ -1,27 +1,24 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
-import '../../../../utill/color_resources.dart';
-import '../../../../utill/custom_themes.dart';
 
 class TextFromFieldWidget extends StatefulWidget {
-  String title;
-  Function validator;
-  Function onChange;
-  TextInputType type;
-  bool isReadOnly;
-  TextEditingController controller;
-  FocusNode focusNode;
-  FocusNode nextNode;
-  bool width;
-  bool height;
-  Color color;
-  bool isRequired;
-  bool obscureText;
+  String? title;
+  final String? Function(String?)? validator;
+  final  Function(String)? onChange;
+  TextInputType? type;
+  bool? isReadOnly;
+  TextEditingController? controller;
+  FocusNode? focusNode;
+  FocusNode? nextNode;
+  bool? width;
+  bool? height;
+  Color? color;
+  bool? isRequired;
+  bool? obscureText;
 
   TextFromFieldWidget(
-      {Key key,
+      {Key? key,
       this.title,
       this.validator,
       this.onChange,
@@ -53,19 +50,19 @@ class _TextFromFieldWidgetState extends State<TextFromFieldWidget> {
         mainAxisAlignment: MainAxisAlignment.start,
         children: [
           Text(
-            widget.title,
+            widget.title!,
             textAlign: TextAlign.right,
-            style: TextStyle(
+            style: const TextStyle(
               color: Color(0xFF212121),
               fontSize: 14,
               fontWeight: FontWeight.w600,
             ),
           ),
-          SizedBox(
+          const SizedBox(
             height: 12,
           ),
-          Container(
-            width: widget.width
+          SizedBox(
+            width: widget.width!
                 ? MediaQuery.of(context).size.width * .38
                 : MediaQuery.of(context).size.width * .9,
             // margin: const EdgeInsets.symmetric(vertical: 12),
@@ -79,30 +76,30 @@ class _TextFromFieldWidgetState extends State<TextFromFieldWidget> {
             child: TextFormField(
               textAlign: TextAlign.start,
               keyboardType: widget.type,
-              style: TextStyle(
+              style: const TextStyle(
                 color: Color(0xCC212121),
                 fontSize: 15,
                 fontWeight: FontWeight.bold,
               ),
-              onChanged: widget.onChange,
-              validator: widget.validator,
-              readOnly: widget.isReadOnly,
+              onChanged: widget.onChange!,
+              validator: widget.validator!,
+              readOnly: widget.isReadOnly!,
               controller: widget.controller,
-              obscureText: widget.obscureText ? isTextObscured : false,
+              obscureText: widget.obscureText! ? isTextObscured : false,
               focusNode: widget.focusNode,
               enableInteractiveSelection: true,
 
-              maxLines: widget.obscureText ? 1 : null,
+              maxLines: widget.obscureText! ? 1 : null,
               onFieldSubmitted: (v) {
                 FocusScope.of(context).requestFocus(widget.nextNode);
               },
 
               decoration: InputDecoration(
-                suffixIcon: widget.obscureText
+                suffixIcon: widget.obscureText!
                     ? IconButton(
                         icon: Icon(
                           eyeIcon,
-                          color: Color(0x66212121),
+                          color: const Color(0x66212121),
                           size: 20,
                         ),
                         onPressed: () {
@@ -114,8 +111,8 @@ class _TextFromFieldWidgetState extends State<TextFromFieldWidget> {
                           });
                         },
                       )
-                    : SizedBox(),
-                contentPadding: EdgeInsets.symmetric(horizontal: 8),
+                    : const SizedBox(),
+                contentPadding: const EdgeInsets.symmetric(horizontal: 8),
                 hintText: widget.title,
                 hintStyle: const TextStyle(
                     color: Color.fromRGBO(33, 47, 62, .61),
@@ -144,7 +141,7 @@ class _TextFromFieldWidgetState extends State<TextFromFieldWidget> {
                   ),
                   borderRadius: BorderRadius.circular(8),
                 ),
-                errorStyle: TextStyle(
+                errorStyle: const TextStyle(
                   fontSize: 10,
                   height: 1,
                 ),
