@@ -8,8 +8,8 @@ import '../../../../../provider/categories_provider.dart';
 import '../../../section_details/section_details_screen.dart';
 
 class ItemSection extends StatefulWidget {
-  CategoriesModel categoriesModel;
-   ItemSection({ this. categoriesModel,key});
+  CategoriesModel? categoriesModel;
+   ItemSection({Key? key,  this. categoriesModel,}) : super(key: key);
 
   @override
   State<ItemSection> createState() => _ItemSectionState();
@@ -21,16 +21,16 @@ class _ItemSectionState extends State<ItemSection> {
     return InkWell(
       onTap: (){
         Provider.of<CategoriesProvider>(context, listen: false)
-            .getSunCategoriesList(context: context , id: widget.categoriesModel.id.toString());
-        AppNavigation.navigateTo(context, SectionDetailsScreen(title:widget.categoriesModel.name.toString() ,));
+            .getSunCategoriesList(context: context , id: widget.categoriesModel!.id.toString());
+        AppNavigation.navigateTo(context, SectionDetailsScreen(title:widget.categoriesModel!.name.toString() ,));
       },
       child: Container(
         height: 50,
         margin: const EdgeInsets.only(bottom: 16),
-        padding: EdgeInsets.symmetric(horizontal: 16),
+        padding: const EdgeInsets.symmetric(horizontal: 16),
         decoration: ShapeDecoration(
           shape: RoundedRectangleBorder(
-            side: BorderSide(width: 1, color: Color(0xFFE1E1E1)),
+            side: const BorderSide(width: 1, color: Color(0xFFE1E1E1)),
             borderRadius: BorderRadius.circular(5),
           ),
         ),
@@ -39,9 +39,9 @@ class _ItemSectionState extends State<ItemSection> {
           crossAxisAlignment: CrossAxisAlignment.center,
           children: [
             Text(
-             widget.categoriesModel.name??'',
+             widget.categoriesModel!.name??'',
               textAlign: TextAlign.center,
-              style: TextStyle(
+              style: const TextStyle(
                 color: Color(0xCC212121),
                 fontSize: 16,
                 fontWeight: FontWeight.w500,
@@ -50,7 +50,7 @@ class _ItemSectionState extends State<ItemSection> {
             ),
             Container(
                 alignment: Alignment.center,
-                child: Icon(Icons.arrow_forward_ios,size: 16,color: Color.fromRGBO(33, 33, 33, 0.4),)),
+                child: const Icon(Icons.arrow_forward_ios,size: 16,color: Color.fromRGBO(33, 33, 33, 0.4),)),
           ],
         ),
       ),
