@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 
-import '../../../../utill/images.dart';
 
 class ItemProfile extends StatefulWidget {
-  final String title;
-  final String icon;
-  final Function onPress;
+  final String? title;
+  final String? icon;
+  final VoidCallback? onPress;
 
-  const ItemProfile({key , this.title, this.onPress,this.icon});
+  const ItemProfile({Key? key , this.title, this.onPress,this.icon}) : super(key: key);
 
   @override
   State<ItemProfile> createState() => _ItemProfileState();
@@ -21,10 +20,10 @@ class _ItemProfileState extends State<ItemProfile> {
       child: Container(
         height: 50,
         margin: const EdgeInsets.only(bottom: 16,left: 20,right: 20),
-        padding: EdgeInsets.symmetric(horizontal: 16),
+        padding: const EdgeInsets.symmetric(horizontal: 16),
         decoration: ShapeDecoration(
           shape: RoundedRectangleBorder(
-            side: BorderSide(width: 1, color: Color(0xFFE1E1E1)),
+            side: const BorderSide(width: 1, color: Color(0xFFE1E1E1)),
             borderRadius: BorderRadius.circular(5),
           ),
         ),
@@ -34,12 +33,12 @@ class _ItemProfileState extends State<ItemProfile> {
           children: [
             Row(
               children: [
-                widget.icon==null?SizedBox():    Image.asset(widget.icon ,width: 16,height: 16,),
-                widget.icon==null?SizedBox():      SizedBox(width: 12,),
+                widget.icon==null?const SizedBox():    Image.asset(widget.icon! ,width: 16,height: 16,),
+                widget.icon==null?const SizedBox():      const SizedBox(width: 12,),
                 Text(
-                widget.title,
+                widget.title!,
                   textAlign: TextAlign.right,
-                  style: TextStyle(
+                  style: const TextStyle(
                     color: Color(0xFF212121),
                     fontSize: 15,
                     fontFamily: 'Tajawal',
@@ -50,7 +49,7 @@ class _ItemProfileState extends State<ItemProfile> {
             ),
             Container(
                 alignment: Alignment.center,
-                child: Icon(Icons.arrow_forward_ios,size: 16,color: Color.fromRGBO(33, 33, 33, 0.4),)),
+                child: const Icon(Icons.arrow_forward_ios,size: 16,color: Color.fromRGBO(33, 33, 33, 0.4),)),
           ],
         ),
       ),
