@@ -40,7 +40,7 @@ class LoggingInterceptor extends InterceptorsWrapper {
 
   @override
   Future onError(DioError err, ErrorInterceptorHandler handler) async {
-    print("ERROR[${err?.response?.statusCode}] => PATH: ${err?.requestOptions?.path}");
+    print("ERROR[${err.response?.statusCode}] => PATH: ${err.requestOptions.path}");
     return super.onError(err, handler);
   }
 }
