@@ -1,4 +1,3 @@
-import 'package:comatecs/view/screen/auth/login/login_screen.dart';
 import 'package:comatecs/view/screen/main_home/main_home_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
@@ -9,7 +8,7 @@ import '../../../../utill/navigation.dart';
 import '../../../base/custom_button.dart';
 
 class SuccessScreen extends StatefulWidget {
-  const SuccessScreen({key});
+  const SuccessScreen({Key? key, }) : super(key: key);
 
   @override
   State<SuccessScreen> createState() => _SuccessScreenState();
@@ -27,7 +26,7 @@ class _SuccessScreenState extends State<SuccessScreen> {
           mainAxisSize: MainAxisSize.max,
           children: [
 
-           SizedBox(width: double.infinity,),
+           const SizedBox(width: double.infinity,),
             SvgPicture.asset(Images.truee),
             Container(
               margin: const EdgeInsets.only(top: 16, bottom: 8),
@@ -44,14 +43,14 @@ class _SuccessScreenState extends State<SuccessScreen> {
             ),
 
             Container(
-                margin: EdgeInsets.only(top: 20),
+                margin: const EdgeInsets.only(top: 20),
                 height: MediaQuery.of(context).size.height*.3,
                 alignment: Alignment.bottomCenter,
                 width: MediaQuery.of(context).size.width * .9,
                 child: CustomButton(
                     btnTxt: 'الصفحة الرئيسية',
                     onTap: () {
-                      AppNavigation.navigateAndFinish(context, MainHomeScreen());
+                      AppNavigation.navigateAndFinish(context, const MainHomeScreen());
                       print('object');
                     })),
 
