@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../../../utill/color_resources.dart';
 
 class CancelOrdersScreen extends StatefulWidget {
-  const CancelOrdersScreen({key});
+  const CancelOrdersScreen({Key? key, }) : super(key: key);
 
   @override
   State<CancelOrdersScreen> createState() => _CancelOrdersScreenState();
@@ -20,11 +20,11 @@ class _CancelOrdersScreenState extends State<CancelOrdersScreen> {
             itemCount: 8,
             itemBuilder: (BuildContext ctx, index) {
               return    Container(
-                margin: EdgeInsets.only(bottom: 16),
-                padding: EdgeInsets.all(16),
+                margin: const EdgeInsets.only(bottom: 16),
+                padding: const EdgeInsets.all(16),
                 decoration: ShapeDecoration(
                   shape: RoundedRectangleBorder(
-                    side: BorderSide(width: 0.50, color: Color(0xFFE1E1E1)),
+                    side: const BorderSide(width: 0.50, color: Color(0xFFE1E1E1)),
                     borderRadius: BorderRadius.circular(5),
                   ),
                 ),
@@ -57,12 +57,12 @@ class _CancelOrdersScreenState extends State<CancelOrdersScreen> {
                         ),
                       ],
                     ),
-                    SizedBox(height: 12,),
-                    Row(
+                    const SizedBox(height: 12,),
+                    const Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         Row(
-                          children: const[
+                          children: [
                             Text(
                               'حالة السداد - ',
                               textAlign: TextAlign.right,
@@ -87,7 +87,7 @@ class _CancelOrdersScreenState extends State<CancelOrdersScreen> {
                             ),
                           ],
                         ),
-                        const Text(
+                        Text(
                           '16.000',
                           textAlign: TextAlign.center,
                           style: TextStyle(
@@ -100,9 +100,9 @@ class _CancelOrdersScreenState extends State<CancelOrdersScreen> {
                         ),
                       ],
                     ),
-                    SizedBox(height: 12,),
-                    Row(
-                      children:const [
+                    const SizedBox(height: 12,),
+                    const Row(
+                      children:[
                         Text(
                           'حالة الطلب - ',
                           textAlign: TextAlign.right,
